Remove unsafe code
After doing some tests the unsafe code showed a 0.2s improvement over a 70mb file, which obviously makes it not worth it.
Also, I saw no difference in performance after testing how badly the while loop ran with and without the if condition inside it, which tells me that the compiler probably optimized it away.